v0.0.4 committed and psuhed from main branch. Gave a good structre and secured using .env
