+++ .gitignore
@@ -71,6 +71,8 @@ typings/
 # dotenv environment variables file
 .env
 .env.test
+# Ignore config files
+config.env
 
 # parcel-bundler cache (https://parceljs.org/)
 .cache
