+++ package.json
@@ -5,6 +5,7 @@
     "cheerio": "^1.0.0-rc.12",
     "colors": "^1.4.0",
     "cors": "^2.8.5",
+    "dotenv": "^16.4.1",
     "express": "^4.18.2",
     "mongoose": "^8.1.1",
     "multer": "^1.4.5-lts.1",
@@ -17,7 +18,7 @@
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error:no test specified\" && exit 1",
-    "start": "nodemon server"
+    "start": "nodemon ./src/server"
   },
   "author": "",
   "license": "ISC",
